refactor(tui): move bottom/status bar render into agent_tui_kit (Ctx slice)
