+++ crates/agent_tui_kit/Cargo.toml
@@ -7,6 +7,7 @@ license = "MIT"
 
 [dependencies]
 tact_protocol = { path = "../protocol" }
+chrono = { workspace = true }
 crossterm = { workspace = true }
 pulldown-cmark = { workspace = true }
 ratatui = { workspace = true }
