+++ crates/agent_tui_kit/src/state/mod.rs
@@ -3,6 +3,7 @@
 //! Phase 3 moves these out of `crates/tui/src/widgets/state` in cluster order.
 //! Each type is pure state + methods — no `App` dependency.
 
+pub mod account;
 pub mod log;
 pub mod log_scroll;
 pub mod mouse_state;
@@ -16,6 +17,7 @@ pub mod thinking;
 pub mod tool_state;
 pub mod ui_types;
 
+pub use account::AccountState;
 pub use log::{LogCoordinator, LogItem, LogItemKind, SystemMsgStyle, log_indent_at};
 pub use log_scroll::LogScroll;
 pub use mouse_state::{LogSelection, MouseState, PopupHitRow, PopupTextHit, TextPosition};
